requirements.txt
Pillow==8.2.0
PySimpleGUI==4.43.0
tk==0.1.0
